(trunk) Merging 3.2 Core enhancements and bug fixes into Trunk.
git-svn-id: https://svn.code.sf.net/p/gridlab-d/code/trunk@5476 54b63ff4-7b51-0410-8bda-9fa96797e505

+++ core/job.cpp
@@ -345,14 +345,14 @@ extern "C" int job(int argc, char *argv[])
 	char mailto[1024];
 	global_getvar("mailto",mailto,sizeof(mailto));
 	
-	unsigned int count = process_dir(global_workdir);
+	unsigned int count = (int)process_dir(global_workdir);
 	if ( count==0 )
 	{
 		output_warning("no models found to process job in workdir '%s'", global_workdir);
 		exit(XC_RUNERR);
 	}
 	
-	int n_procs = global_threadcount;
+	unsigned int n_procs = global_threadcount;
 	if ( n_procs==0 ) n_procs = processor_count();
 	pthread_t *pid = new pthread_t[n_procs];
 	output_debug("starting job with cmdargs '%s' using %d threads", job_cmdargs, n_procs);
